requirements.txt
streamlit>=1.37,<2
pandas>=2.2,<3
numpy>=1.26,<3
scikit-learn>=1.4,<2
umap-learn>=0.5.6,<0.6
plotly>=5.22,<6
pyarrow>=15,<18
Pillow>=10,<12
